fix and refactor price calculation

+++ backend/app/Infrastructure/Map/DistanceCalculation/DistanceCalculatorInterface.php
@@ -6,5 +6,5 @@ use Fhaculty\Graph\Vertex;
 
 interface DistanceCalculatorInterface
 {
-    public function calculateDistance(Vertex $vertex1, Vertex $vertex2): float;
+    public function calculateDistance(Vertex $vertex1, Vertex $vertex2): int;
 }
